+++ fe/lib/components/SalaryDetails/styles.css.ts
