--- fe/lib/components/SalaryDetails/v2/README.md
@@ -1,98 +1,0 @@
-# SalaryDetailsV2
-
-A Salary Details component that contains information on the salary type, range and description.
-
-## Installation
-
-```shell
-$ yarn add wingman-fe
-```
-
-### Validation
-
-The Salary Details component has some basic UI validation rules:
-
-- The minimum pay must be greater than 0.
-- The maximum pay must be greater than the minimum pay.
-
-### Usage
-
-```javascript
-import React, { useState } from 'react';
-import { SalaryDetailsV2 } from 'wingman-fe';
-
-const buildRemunerationPackageInput = (remuneration) => ({
-  basisCode: remuneration.basisCode,
-  ranges: [{
-    intervalCode: remuneration.intervalCode,
-    minimumAmount: {
-      currency: remuneration.currency,
-      value: Number(remuneration.minimumAmount),
-    },
-    maximumAmount: remuneration.maximumAmount ?? {
-      currency: remuneration.currency,
-      value: Number(remuneration.maximumAmount),
-    },
-  }],
-  descriptions: remuneration.description ? [remuneration.description] : [],
-})
-
-const PostingForm = () => {
-  const [remuneration, setRemuneration] = useState({
-    basisCode: '',
-    currency: 'AUD',
-    description: '',
-    intervalCode: '',
-    maximumAmount: null,
-    minimumAmount: '',
-    payType: '',
-  });
-
-  const onBlur = (item) => {
-    switch (item.key) {
-      case 'payType':
-        setRemuneration({
-          ...remuneration,
-          basis: item.basis,
-          interval: item.interval,
-          payType: item.payType,
-        });
-        break;
-
-      case 'minimumAmount':
-        setRemuneration({
-          ...remuneration,
-          minimumAmount: item.amount,
-        });
-        break;
-
-      case 'maximumAmount':
-        setRemuneration({
-          ...remuneration,
-          maximumAmount: item.amount,
-        });
-        break;
-
-      case 'description':
-        setRemuneration({
-          ...remuneration,
-          description: item.description,
-        });
-        break;
-
-      case 'currency':
-        setRemuneration({
-          ...remuneration,
-          currency: item.currency,
-        });
-        break;
-    }
-  }
-
-  return (
-    <div>
-      <SalaryDetailsV2 initialCurrency="AUD" onBlur={onBlur}>
-    </div>
-  );
-};
-```
