--- fe/lib/components/SalaryDetails/v2/SalaryDetails.mock.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-
-import { MockComponentActions } from '../../../private/MockComponentActions/MockComponentActions';
-
-import { SalaryDetails, type SalaryDetailsProps } from './SalaryDetails';
-
-interface Props extends SalaryDetailsProps {
-  showStorybookAction?: boolean;
-}
-
-export const MockSalaryDetails = ({ showStorybookAction, ...props }: Props) => (
-  <MockComponentActions
-    space="large"
-    showStorybookAction={showStorybookAction}
-    storybookPath="/story/job-posting-salary-details--salary-details"
-    sourcePath="lib/components/SalaryDetailsV2"
-  >
-    <SalaryDetails {...props} />
-  </MockComponentActions>
-);
